+++ examples/rust/src/routes/index.server.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import {fib} from '../rust.server';
+
+export default function Home() {
+  return (
+    <>
+      <p>Hello: {fib(30)}</p>
+    </>
+  );
+}