+++ packages/hydrogen/src/foundation/useQuery/hooks.ts
@@ -3,7 +3,6 @@ import {
   getLoggerWithContext,
   collectQueryCacheControlHeaders,
   collectQueryTimings,
-  logCacheApiStatus,
 } from '../../utilities/log';
 import {
   deleteItemFromCache,
@@ -11,10 +10,9 @@ import {
   getItemFromCache,
   isStale,
   setItemInCache,
-} from '../../framework/cache';
-import {hashKey} from '../../utilities/hash';
-import {runDelayedFunction} from '../../framework/runtime';
+} from '../Cache/cache-sub-request';
 import {useRequestCacheData, useServerRequest} from '../ServerRequestProvider';
+import {CacheSeconds} from '../Cache/strategies';
 
 export interface HydrogenUseQueryOptions {
   /** The [caching strategy](https://shopify.dev/custom-storefronts/hydrogen/framework/cache#caching-strategies) to help you
@@ -90,7 +88,6 @@ function cachedQueryFnBuilder<T>(
     // to prevent losing the current React cycle.
     const request = useServerRequest();
     const log = getLoggerWithContext(request);
-    const hashedKey = hashKey(key);
 
     const cacheResponse = await getItemFromCache(key);
 
@@ -110,28 +107,38 @@ function cachedQueryFnBuilder<T>(
       /**
        * Important: Do this async
        */
-      if (isStale(response, resolvedQueryOptions?.cache)) {
-        logCacheApiStatus('STALE', hashedKey);
-        const lockKey = `lock-${key}`;
+      if (isStale(key, response)) {
+        const lockKey = ['lock', ...(typeof key === 'string' ? [key] : key)];
 
-        runDelayedFunction(async () => {
-          logCacheApiStatus('UPDATING', hashedKey);
-          const lockExists = await getItemFromCache(lockKey);
-          if (lockExists) return;
+        // Run revalidation asynchronously
+        const revalidatingPromise = getItemFromCache(lockKey).then(
+          async (lockExists) => {
+            if (lockExists) return;
 
-          await setItemInCache(lockKey, true);
-          try {
-            const output = await generateNewOutput();
+            await setItemInCache(
+              lockKey,
+              true,
+              CacheSeconds({
+                maxAge: 10,
+              })
+            );
 
-            if (shouldCacheResponse(output)) {
-              await setItemInCache(key, output, resolvedQueryOptions?.cache);
+            try {
+              const output = await generateNewOutput();
+
+              if (shouldCacheResponse(output)) {
+                await setItemInCache(key, output, resolvedQueryOptions?.cache);
+              }
+            } catch (e: any) {
+              log.error(`Error generating async response: ${e.message}`);
+            } finally {
+              await deleteItemFromCache(lockKey);
             }
-          } catch (e: any) {
-            log.error(`Error generating async response: ${e.message}`);
-          } finally {
-            await deleteItemFromCache(lockKey);
           }
-        });
+        );
+
+        // Asynchronously wait for it in workers
+        request.ctx.runtime?.waitUntil?.(revalidatingPromise);
       }
 
       return output;
@@ -143,9 +150,13 @@ function cachedQueryFnBuilder<T>(
      * Important: Do this async
      */
     if (shouldCacheResponse(newOutput)) {
-      runDelayedFunction(() =>
-        setItemInCache(key, newOutput, resolvedQueryOptions?.cache)
+      const setItemInCachePromise = setItemInCache(
+        key,
+        newOutput,
+        resolvedQueryOptions?.cache
       );
+
+      request.ctx.runtime?.waitUntil?.(setItemInCachePromise);
     }
 
     collectQueryCacheControlHeaders(