+++ packages/hydrogen/src/hooks/useShopQuery/hooks.ts
@@ -1,9 +1,7 @@
 import {useShop} from '../../foundation/useShop';
 import {getLoggerWithContext} from '../../utilities/log';
-import {ASTNode} from 'graphql';
 import type {CachingStrategy, PreloadOptions} from '../../types';
 import {graphqlRequestBody} from '../../utilities';
-import {getConfig} from '../../framework/config';
 import {useServerRequest} from '../../foundation/ServerRequestProvider';
 import {injectGraphQLTracker} from '../../utilities/graphql-tracker';
 import {sendMessageToClient} from '../../utilities/devtools';
@@ -42,7 +40,7 @@ export function useShopQuery<T>({
   /** A string of the GraphQL query.
    * If no query is provided, useShopQuery will make no calls to the Storefront API.
    */
-  query?: ASTNode | string;
+  query?: string;
   /** An object of the variables for the GraphQL query. */
   variables?: Record<string, any>;
   /** The [caching strategy](https://shopify.dev/custom-storefronts/hydrogen/framework/cache#caching-strategies) to
@@ -104,7 +102,7 @@ export function useShopQuery<T>({
     log.error(errorMessage);
     log.error(useQueryError);
 
-    if (getConfig().dev) {
+    if (__HYDROGEN_DEV__ && !__HYDROGEN_TEST__) {
       throw new Error(errorMessage);
     } else {
       // in non-dev environments, we probably don't want super-detailed error messages for the user
@@ -122,7 +120,7 @@ export function useShopQuery<T>({
     const errors = Array.isArray(data.errors) ? data.errors : [data.errors];
 
     for (const error of errors) {
-      if (getConfig().dev) {
+      if (__HYDROGEN_DEV__ && !__HYDROGEN_TEST__) {
         throw new Error(error.message);
       } else {
         log.error('GraphQL Error', error);
@@ -132,10 +130,9 @@ export function useShopQuery<T>({
   }
 
   if (
-    __DEV__ &&
+    __HYDROGEN_DEV__ &&
     log.options().showUnusedQueryProperties &&
     query &&
-    typeof query !== 'string' &&
     data?.data
   ) {
     const fileLine = new Error('').stack