+++ packages/hydrogen/src/utilities/apiRoutes.ts
@@ -1,14 +1,12 @@
 import {
-  HydrogenConfig,
-  HydrogenConfigRoutes,
+  InlineHydrogenConfig,
+  ResolvedHydrogenRoutes,
   ImportGlobEagerOutput,
 } from '../types';
 import {matchPath} from './matchPath';
 import {getLoggerWithContext, logServerResponse} from '../utilities/log/';
-import type {ServerComponentRequest} from '../framework/Hydration/ServerComponentRequest.server';
-import type {ASTNode} from 'graphql';
+import type {HydrogenRequest} from '../foundation/HydrogenRequest/HydrogenRequest.server';
 import {fetchBuilder, graphqlRequestBody} from './fetch';
-import {findRoutePrefix} from './findRoutePrefix';
 import {getStorefrontApiRequestHeaders} from './storefrontApi';
 import {
   emptySessionImplementation,
@@ -70,24 +68,21 @@ export function extractPathFromRoutesKey(
   return path;
 }
 
-export function getApiRoutes(
-  rawRoutes: HydrogenConfigRoutes
-): Array<HydrogenApiRoute> {
-  const routes = (rawRoutes.files ?? rawRoutes) as ImportGlobEagerOutput;
-  const topLevelPath = (rawRoutes.basePath ?? '*') as string;
-  const dirPrefix = rawRoutes.dirPrefix as string | undefined;
-
+export function getApiRoutes({
+  files: routes,
+  basePath: topLevelPath = '',
+  dirPrefix = '',
+}: Partial<ResolvedHydrogenRoutes>): Array<HydrogenApiRoute> {
   if (!routes || memoizedRawRoutes === routes) return memoizedApiRoutes;
 
   const topLevelPrefix = topLevelPath.replace('*', '').replace(/\/$/, '');
 
   const keys = Object.keys(routes);
-  const commonRoutePrefix = dirPrefix ?? findRoutePrefix(keys);
 
   const apiRoutes = keys
     .filter((key) => routes[key].api)
     .map((key) => {
-      const path = extractPathFromRoutesKey(key, commonRoutePrefix);
+      const path = extractPathFromRoutesKey(key, dirPrefix);
 
       /**
        * Catch-all routes [...handle].jsx don't need an exact match
@@ -145,14 +140,14 @@ interface QueryShopArgs {
   /** A string of the GraphQL query.
    * If no query is provided, then the `useShopQuery` makes no calls to the Storefront API.
    */
-  query: ASTNode | string;
+  query: string;
   /** An object of the variables for the GraphQL query. */
   variables?: Record<string, any>;
 }
 
 function queryShopBuilder(
-  shopifyConfigGetter: HydrogenConfig['shopify'],
-  request: ServerComponentRequest
+  shopifyConfigGetter: InlineHydrogenConfig['shopify'],
+  request: HydrogenRequest
 ) {
   return async function queryShop<T>({
     query,
@@ -194,9 +189,9 @@ function queryShopBuilder(
 }
 
 export async function renderApiRoute(
-  request: ServerComponentRequest,
+  request: HydrogenRequest,
   route: ApiRouteMatch,
-  shopifyConfig: HydrogenConfig['shopify'],
+  shopifyConfig: InlineHydrogenConfig['shopify'],
   session?: SessionStorageAdapter
 ): Promise<Response | Request> {
   let response;
@@ -250,7 +245,7 @@ export async function renderApiRoute(
 
   logServerResponse(
     'api',
-    request as ServerComponentRequest,
+    request as HydrogenRequest,
     (response as Response).status ?? 200
   );