+++ packages/playground/server-components/hydrogen.config.js
@@ -2,7 +2,6 @@ import {CookieSessionStorage} from '@shopify/hydrogen';
 import {defineConfig} from '@shopify/hydrogen/config';
 
 export default defineConfig({
-  routes: import.meta.globEager('./src/routes/**/*.server.[jt](s|sx)'),
   shopify: {
     defaultLocale: 'en-us',
     storeDomain: 'hydrogen-preview.myshopify.com',
@@ -12,4 +11,8 @@ export default defineConfig({
   session: CookieSessionStorage('__session', {
     expires: new Date(1749343178614),
   }),
+  logger: {
+    trace() {},
+    debug() {},
+  },
 });