+++ packages/playground/server-components/src/components/ClientParams.client.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {useRouteParams} from '@shopify/hydrogen/client';
+import {useRouteParams} from '@shopify/hydrogen';
 
 export default function ClientParams() {
   const {handle} = useRouteParams();