+++ templates/demo-store/src/routes/account/register.server.jsx
@@ -1,0 +1,81 @@
+import {NoStore, Seo, gql} from '@shopify/hydrogen';
+
+import {getApiErrorMessage} from '../../components/utilities/api.helper';
+import Layout from '../../components/Layout.server';
+import AccountCreateForm from '../../components/account/AccountCreateForm.client';
+
+export default function Register({response}) {
+  response.cache(NoStore());
+
+  return (
+    <Layout>
+      <Seo type="noindex" data={{title: 'Register'}} />
+      <AccountCreateForm />
+    </Layout>
+  );
+}
+
+export async function api(request, {queryShop}) {
+  const jsonBody = await request.json();
+
+  if (
+    !jsonBody.email ||
+    jsonBody.email === '' ||
+    !jsonBody.password ||
+    jsonBody.password === ''
+  ) {
+    return new Response(
+      JSON.stringify({error: 'Email and password are required'}),
+      {status: 400},
+    );
+  }
+
+  const {data, errors} = await queryShop({
+    query: MUTATION,
+    variables: {
+      input: {
+        email: jsonBody.email,
+        password: jsonBody.password,
+        firstName: jsonBody.firstName,
+        lastName: jsonBody.lastName,
+      },
+    },
+    cache: NoStore(),
+  });
+
+  const errorMessage = getApiErrorMessage('customerCreate', data, errors);
+
+  if (
+    !errorMessage &&
+    data &&
+    data.customerCreate &&
+    data.customerCreate.customer &&
+    data.customerCreate.customer.id
+  ) {
+    return new Response(null, {
+      status: 200,
+    });
+  } else {
+    return new Response(
+      JSON.stringify({
+        error: errorMessage ?? 'Unknown error',
+      }),
+      {status: 401},
+    );
+  }
+}
+
+const MUTATION = gql`
+  mutation customerCreate($input: CustomerCreateInput!) {
+    customerCreate(input: $input) {
+      customer {
+        id
+      }
+      customerUserErrors {
+        code
+        field
+        message
+      }
+    }
+  }
+`;